[api-auth-v1] refactor: User Package 구조 변경 (#40)
