+++ api/auth-v1/src/main/java/kr/joberchip/auth/v1/user/dto/UserRequest.java
@@ -1,4 +1,4 @@
-package kr.joberchip.auth.v1.dto;
+package kr.joberchip.auth.v1.user.dto;
 
 import kr.joberchip.core.user.User;
 import lombok.Getter;
